feat(steps): Break out 'tag' subcommand
This fills in any of the tags not currently set.
